6-sum_mixed_list.py update
Change order of type annotation

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -3,7 +3,7 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """Sum elements in list of floats and ints."""
     sum: float = 0
     for n in mxd_lst:
